separated some code
