+++ src/error_msg.rs
@@ -1,0 +1,3 @@
+pub fn expect(expecting: &str, got: &str) -> String {
+    format!("expecting {}, got {}", expecting, got)
+}
